Testsuite fixups
Conflicts:
	gcc/testsuite/c-c++-common/Wlogical-op-1.c
	gcc/testsuite/c-c++-common/builtin-shuffle-1.c
	gcc/testsuite/c-c++-common/pr60226.c
	gcc/testsuite/c-c++-common/pr68657-1.c
	gcc/testsuite/c-c++-common/pr68657-2.c
	gcc/testsuite/c-c++-common/pr68833-1.c
	gcc/testsuite/g++.dg/cpp0x/constexpr-70001-3.C
	gcc/testsuite/g++.dg/cpp0x/constexpr-70001-4.C
	gcc/testsuite/g++.dg/lto/20080908-1_0.C
	gcc/testsuite/g++.dg/lto/pr69589_0.C
	gcc/testsuite/g++.dg/warn/Wtype-limits-Wextra.C
	gcc/testsuite/g++.dg/warn/Wtype-limits.C
	gcc/testsuite/g++.old-deja/g++.jason/thunk3.C
	gcc/testsuite/gcc.c-torture/compile/pr37669.c
	gcc/testsuite/gcc.c-torture/compile/pr69102.c
	gcc/testsuite/gcc.c-torture/execute/bitfld-6.c
	gcc/testsuite/gcc.c-torture/execute/bitfld-7.c
	gcc/testsuite/gcc.c-torture/execute/pr38151.c
	gcc/testsuite/gcc.c-torture/execute/pr66556.c
	gcc/testsuite/gcc.c-torture/execute/pr68648.c
	gcc/testsuite/gcc.c-torture/execute/pr70602.c
	gcc/testsuite/gcc.c-torture/execute/pr71083.c
	gcc/testsuite/gcc.dg/Warray-bounds-11.c
	gcc/testsuite/gcc.dg/Wno-frame-address.c
	gcc/testsuite/gcc.dg/enum-mode-1.c
	gcc/testsuite/gcc.dg/graphite/scop-22.c
	gcc/testsuite/gcc.dg/graphite/scop-3.c
	gcc/testsuite/gcc.dg/graphite/scop-dsyr2k.c
	gcc/testsuite/gcc.dg/graphite/scop-dsyrk.c
	gcc/testsuite/gcc.dg/graphite/scop-mvt.c
	gcc/testsuite/gcc.dg/graphite/scop-sor.c
	gcc/testsuite/gcc.dg/ifcvt-4.c
	gcc/testsuite/gcc.dg/init-excess-2.c
	gcc/testsuite/gcc.dg/ipa/propalign-4.c
	gcc/testsuite/gcc.dg/pr59471.c
	gcc/testsuite/gcc.dg/pr62090-2.c
	gcc/testsuite/gcc.dg/pr69634.c
	gcc/testsuite/gcc.dg/pr69973.c
	gcc/testsuite/gcc.dg/sso-6.c
	gcc/testsuite/gcc.dg/torture/pr68067-1.c
	gcc/testsuite/gcc.dg/torture/pr68067-2.c
	gcc/testsuite/gcc.dg/torture/pr69352.c
	gcc/testsuite/gcc.dg/torture/pr69714.c
	gcc/testsuite/gcc.dg/torture/pr69771.c
	gcc/testsuite/gcc.dg/torture/pr69941.c
	gcc/testsuite/gcc.dg/torture/pr70421.c
	gcc/testsuite/gcc.dg/torture/pr70542.c
	gcc/testsuite/gcc.dg/tree-ssa/pr64130.c
	gcc/testsuite/gcc.dg/tree-ssa/pr65136.c
	gcc/testsuite/gcc.dg/tree-ssa/pr66449.c
	gcc/testsuite/gcc.dg/tree-ssa/pr66726.c
	gcc/testsuite/gcc.dg/tree-ssa/pr68529-3.c
	gcc/testsuite/gcc.dg/tree-ssa/pr70919.c
	gcc/testsuite/gcc.dg/tree-ssa/scev-8.c
	gcc/testsuite/gcc.dg/tree-ssa/sra-20.c
	gcc/testsuite/gcc.dg/tree-ssa/ssa-dom-thread-8.c
	gcc/testsuite/gcc.dg/uninit-19.c
	libstdc++-v3/testsuite/20_util/tuple/creation_functions/constexpr.cc
	libstdc++-v3/testsuite/25_algorithms/nth_element/58800.cc
	libstdc++-v3/testsuite/26_numerics/random/linear_congruential_engine/requirements/constexpr_data.cc
	libstdc++-v3/testsuite/26_numerics/random/linear_congruential_engine/requirements/constexpr_functions.cc
	libstdc++-v3/testsuite/27_io/basic_istream/getline/wchar_t/5.cc
	libstdc++-v3/testsuite/27_io/ios_base/types/fmtflags/case_label.cc
	libstdc++-v3/testsuite/27_io/ios_base/types/iostate/case_label.cc
	libstdc++-v3/testsuite/27_io/ios_base/types/openmode/case_label.cc

+++ gcc/testsuite/gcc.dg/graphite/pr42205-1.c
@@ -1,7 +1,10 @@
 /* { dg-require-effective-target int32plus } */
 /* { dg-options "-O1 -ffast-math -floop-interchange" } */
+/* { dg-require-effective-target stdint_types } */
 
-int adler32(int adler, char *buf, int n)
+#include <stdint.h>
+
+int_least32_t adler32(int adler, char *buf, int n)
 {
   int sum = 0;
   do {
@@ -14,5 +17,5 @@ int adler32(int adler, char *buf, int n)
      adler += buf[3];
      sum += adler;
   } while (--n);
-  return adler | (sum << 16);
+  return adler | ((int_least32_t)sum << 16);
 }
