+++ gcc/testsuite/gcc.dg/pr66299-1.c
@@ -1,7 +1,9 @@
 /* PR tree-optimization/66299 */
-/* { dg-do run } */
+/* { dg-do run { target { stdint_types } } } */
 /* { dg-options "-fdump-tree-original" } */
 /* { dg-require-effective-target int32plus } */
+
+#include <stdint.h>
 
 void
 test1 (int x)
@@ -72,11 +74,11 @@ test4 (unsigned int x)
 }
 
 void
-test5 (int x)
+test5 (int_least32_t x)
 {
-  if ((0 << x) == 1
-      || (0 << x) != 0
-      || (0x8001U << x) != 0x20000U)
+  if (((int_least32_t)0 << x) == 1
+      || ((int_least32_t)0 << x) != 0
+      || ((int_least32_t)0x8001U << x) != 0x20000U)
     __builtin_abort ();
 }
